% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{callACRs}
\alias{callACRs}
\title{callACRs}
\usage{
callACRs(
  obj,
  genomesize = 1.6e+09,
  shift = -50,
  extsize = 100,
  output = "bulk_peaks",
  tempdir = "./macs2_temp",
  verbose = T,
  fdr = 0.05
)
}
\arguments{
\item{obj}{Object output from loadBEDandGenomeData. Required.}

\item{genomesize}{Effective genome size parameter for MACS2 (defaults to 1.6e9 for Zea mays).}

\item{shift}{Shift the start coordinates upstream. Defaults to -50 bp.}

\item{extsize}{Extends the fragment length downstream. Defaults to 100 bp.}

\item{output}{Output file names for MACS2. Defaults to 'bulk_peaks'.}

\item{tempdir}{Directory name to store the output of MACS2. Defaults to "./macs2_temp".}

\item{verbose}{Logical. Whether to print information.}

\item{fdr}{Float. Significance cut-off for peak calling. Defaults to q-value 0.05.}
}
\description{
This function calls ACRs using the bulk Tn5 integration sites with MACS2
}
