% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{callClusters}
\alias{callClusters}
\title{Cluster cells using Louvain, Leiden or other graph-based methods implemented by Seurat}
\usage{
callClusters(
  obj,
  res = 0.4,
  k.near = 30,
  clustOB = "svd",
  cname = "LouvainClusters",
  min.reads = 50000,
  m.clst = 25,
  threshold = 5,
  umap1 = "umap1",
  umap2 = "umap2",
  e.thresh = 3,
  cleanCluster = T,
  cl.method = 1,
  svd_slotName = "PCA",
  umap_slotName = "UMAP",
  cluster_slotName = "Clusters",
  verbose = FALSE
)
}
\arguments{
\item{obj}{list, object containing 'PCA', 'UMAP', 'counts' and 'meta'. Required.}

\item{res}{numeric, resolution for Louvain/Leiden clustering. Typical values range from 0 to 2.
Defaults to 0.8.}

\item{k.near}{numeric, number of nearest neighbors for graph building. Defaults to 20.}

\item{clustOB}{character, which embedding to use for clustering. Can be one of c("umap", "svd").
It is highly recommended to use 'svd' for graph-based clustering. Defaults to "svd".}

\item{cname}{character, column name in meta cluster IDs. Defaults to "LouvainClusters".}

\item{min.reads}{numeric, minimum number of aggregated nSites (number of accessible peaks per
cell) to consider a cluster as valid. Note: that this does not consider the total number of reads
in each cell. Rather, we take the sum of column sums (colSum) for cells within a single cluster.
Defaults to 5e4.}

\item{m.clst}{numeric, minimum number of cells to consider a cluster as valid. Defaults to 25.}

\item{threshold}{numeric, Z-score threshold to filter out cells greater than X distance to other
cells on average. Uses a knn to find the top 50 nearest cells. See 'filterSingle' for more
details.}

\item{umap1}{character, column name of UMAP first component. Only important for sub-clustering.}

\item{umap2}{character, column name of UMAP second component. Only important for sub-clustering.}

\item{e.thresh}{numeric, Z-score threshold to filter out cells failing to co-localize with other
cells apart of the same cluster. Set to 1e6 to skip cluster refining. This step may also be
skipped by setting cleanCluster to FALSE. Default is 2.}

\item{cleanCluster}{logical, whether or not to implement cluster refinement by removing cells
that do not co-localize the the majority of cells apart of the same cluster. Uses the UMAP
embedding for distance estimates.}

\item{cl.method}{numeric, graph clustering algorithm. Numeric values range from 1-4, see
Seurat::FindClusters for more details.}

\item{svd_slotName}{character, character string for the SVD slot to use for clustering. Defaults
to "PCA".}

\item{umap_slotName}{character, character string for the UMAP slot to use for analysis. Defaults
to "UMAP.}

\item{cluster_slotName}{character, character string for naming the results of callClusters.
Defaults to "Clusters".}

\item{verbose}{logical. Defaults to FALSE.}

\item{key}{character, name of UMAP key when creating Seurat object. Defaults to "UMAP_".}
}
\description{
Cluster cells using Louvain, Leiden or other graph-based methods implemented by Seurat
}
