% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cleanData}
\alias{cleanData}
\title{Filter sparse matrix}
\usage{
cleanData(
  obj,
  preclusterID = NULL,
  min.p = 0.01,
  min.t = 0.001,
  min.c = 100,
  max.t = 0.005,
  min.cells.precluster = 10,
  verbose = F
)
}
\arguments{
\item{obj}{list (counts and meta) output from loadSparseData}

\item{preclusterID}{character column name in meta data for LSI or other clustering results to
condition ACR peak frequency. Sets the minimum frequency of accessibility across cells within a
single cluster, where sites that pass this threshold in at least one cluster are retained.
Defaults to NULL.}

\item{min.p}{numeric, minimum frequency of accessible cells per cluster (see preclusterID).
Sites above this threshold in at least one cluster are retained. Acceptable values range from 0
to 1. Defaults to 0.01.}

\item{min.t}{filters ACRs below the specified frequency of accessible cells across all cells.
Acceptable values range from 0 to 1.}

\item{min.c}{minimum number of accessible ACRs per cell. Defaults to 100.}

\item{max.t}{remove top X\% of ACRs by frequency across cells (remove constitutive accessible
regions). Defaults to 0.005}

\item{min.cells.precluster}{minimum number cells from precluster (preclusterID) to include
cluster for thresholds set by min.p. Defaults to 10}

\item{verbose}{Defaults to FALSE}
}
\description{
This function filters the cell x peak matrix based on pre-clustering results, a minimum peak
usage frequency, and number of accessible peaks per cell
}
