% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccessibility.R
\name{coAccess}
\alias{coAccess}
\title{coAccess}
\usage{
coAccess(
  obj,
  genome = NULL,
  byGroup = FALSE,
  groupID = "LouvainClusters",
  eFDR = FALSE,
  fdr_thresh = 0.05,
  win_size = 5e+05,
  sample_num = 100,
  k = 30,
  conn_slotName = "coACRs",
  nthreads = 1,
  verbose = FALSE
)
}
\arguments{
\item{obj}{Socrates object. Required.}

\item{genome}{path to genome file. See 'inst/extdata/genome' for an example. Required.}

\item{byGroup}{logical. Whether to run co-accessibility tests for each factor in 'groupID'.
Defaults to False.}

\item{groupID}{character. Column name in obj$meta for partitioning during co-accessibility
analysis. Defaults to "LouvainClusters".}

\item{eFDR}{logical. Whether or not to use empirical FDR to filter co-accessible ACRs. Defaults
to FALSE. Setting eFDR to 'TRUE' doubles memory usage.}

\item{fdr_thresh}{float. fdr threshold when using empirical FDR. Defaults to 0.05.}

\item{win_size}{numeric. Genomic distance to constrain co-accessible ACRs. Defaults to 500000.}

\item{sample_num}{numeric. Number of random windows to sample for determining distance_parameter
during 'run_cicero'. See cicero for more details.}

\item{k}{numeric. Number of nearest neighbors for aggregating cells. Defaults to 30.}

\item{conn_slotName}{character. Slot name for depositing co-accessible ACRs. Defaults to coACRs.}

\item{nthreads}{numeric. Number of threads to use for estimating co-ACRs for each cluster or
celltype. Defaults to 1.}

\item{verbose}{logical. Defaults to FALSE.}
}
\description{
This function identifies co-accessible ACRs within clusters or across the entire data set
}
