% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convertSparseData}
\alias{convertSparseData}
\title{convertSparseData}
\usage{
convertSparseData(obj, verbose = F)
}
\arguments{
\item{obj}{Object output by generateMatrix.}

\item{verbose}{logical. Defaults to FALSE.}
}
\description{
This function loads the sparse matrix and meta data and returns a list - counts & meta
}
