% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{countRemoveOrganelle}
\alias{countRemoveOrganelle}
\title{countRemoveOrganelle}
\usage{
countRemoveOrganelle(obj, org_scaffolds = NULL, remove_reads = FALSE)
}
\arguments{
\item{obj}{Object output from loadBEDandGenomeData. Required.}

\item{org_scaffolds}{vector of organelle scaffold names}

\item{remove_reads}{Logical of whether to remove reads from the BED file or not. Defaults to TRUE}
}
\description{
This functions takes in an the Tn5 bed file, as well as obj vector
which contains organelle scaffolds, and identifies Tn5 integrations events which occured
in organelles. These reads are then assigned to the object slot PtMT. Additional parameters
allow for the removal of these reads so they don't interfere with ACR calling downstream.
}
