% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_analysis.R
\name{detectDoublets}
\alias{detectDoublets}
\title{detectDoublets}
\usage{
detectDoublets(
  obj = NULL,
  nTrials = 5,
  nSample = 1000,
  k = 10,
  threads = 1,
  rdMethod = NULL,
  svd_slotName = "PCA"
)
}
\arguments{
\item{obj}{Socrates object. For best results, use downstream of the function cleanCells prior to
normalization.}

\item{nTrials}{Numeric. Number of times nSample cells are simulated.}

\item{nSample}{Numeric. Number of synthetic doublets to create per trial.}

\item{k}{Numeric. Number of nearest neighbors to search for when estimating doublet enrichment via knn
from the narbor package.}

\item{threads}{Numeric. Number of threads to use for mclapply from the parallel package.}

\item{rdMethod}{Character. Either "SVD" or "NMF" for dimensionality reduction used previously. Defaults
to NULL, using the method specified in obj$rdMethod.}

\item{svd_slotName}{Character. Slot name containing reduced dimensions from reduceDims function.
Default is PCA.}
}
\description{
This function estimates doublet likelihood via synthetic doublet creation and enrichment
analysis. The underlying code is heavily adapted from ArchR described by Granja and Corces et al.
If you use detectDoublets in your analysis, please cite the original ArchR paper.
}
