% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{filtDistClst}
\alias{filtDistClst}
\title{Filter singleton cells from the UMAP embedding by distance per cluster}
\usage{
filtDistClst(b, umap1 = "umap1", umap2 = "umap2", threshold2 = 2)
}
\arguments{
\item{b}{meta data after clustering with Seurat}

\item{threshold2}{numeric, Z-score threshold to remove cells greater than X standard standard
deviations from other cells.}

\item{m1}{character, if type=="umap", then set to the column name of first component ("umap1").
Only used when running sub-clustering (umap embedding names for subset and regenerated UMAP
embeddings of a subset of cells).}

\item{m2}{character, if type=="umap", then set to the column name of second component ("umap2").
#' Only used when running sub-clustering (umap embedding names for subset and regenerated UMAP
embeddings of a subset of cells).}
}
\description{
This function removes cells from the UMAP embedding that are not well supported by other cells.
Can be run on a per cluster basis to filter cells that exceed a heuristic threshold of distances
to most other cells in the same cluster.
}
