% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_analysis.R
\name{filterDoublets}
\alias{filterDoublets}
\title{filterDoublets}
\usage{
filterDoublets(obj = NULL, filterRatio = 1.5, embedding = "UMAP", verbose = T)
}
\arguments{
\item{obj}{Socrates object. For best results, use downstream of the function cleanCells prior to
normalization.}

\item{filterRatio}{Numeric.}

\item{embedding}{Character.}

\item{verbose}{Boolean.}
}
\description{
This function estimates doublet likelihood via synthetic doublet creation and enrichment
analysis. The underlying code is heavily adapted from ArchR described by Granja and Corces et al.
If you use detectDoublets in your analysis, please cite the original ArchR paper.
}
