% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{filterSingle}
\alias{filterSingle}
\title{Filter singleton cells from the UMAP embedding}
\usage{
filterSingle(
  pro,
  k = 50,
  threshold = 3,
  type = "umap",
  m1 = "umap1",
  m2 = "umap2"
)
}
\arguments{
\item{pro}{data.frame containing UMAP coordinates m1, and m2. See m1, m2 arguments for details.}

\item{k}{numeric, number of nearest neighbors. Defaults to 50.}

\item{threshold}{numeric, Z-score threshold to remove cells greater than X standard standard
deviations from other cells.}

\item{type}{character, which embedding to use. Must be one of "umap", "pca", or "tsne".}

\item{m1}{character, if type=="umap", then set to the column name of first component ("umap1").
Only used when running sub-clustering (umap embedding names for subset and regenerated UMAP
embeddings of a subset of cells).}

\item{m2}{character, if type=="umap", then set to the column name of second component ("umap2").
#' Only used when running sub-clustering (umap embedding names for subset and regenerated UMAP
embeddings of a subset of cells).}
}
\description{
This function removes cells from the UMAP embedding that are not well supported by other cells.
Can be run on a per cluster basis to filter cells that exceed a heuristic threshold of distances
to most other cells in the same cluster.
}
