% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{findCells}
\alias{findCells}
\title{findCells}
\usage{
findCells(
  obj,
  set.tn5.cutoff = NULL,
  min.cells = 1000,
  max.cells = 15000,
  min.tn5 = 1000,
  filt.org = F,
  org.filter.thresh = 0.8,
  filt.tss = T,
  tss.min.freq = 0.2,
  tss.z.thresh = 2,
  filt.frip = T,
  frip.min.freq = 0.2,
  frip.z.thresh = 2,
  doplot = F,
  prefix = NULL
)
}
\arguments{
\item{obj}{Object output from buildMetaData. Required.}

\item{set.tn5.cutoff}{Override spline fitting to set minimum tn5 count per cell.
Defaults to NULL.}

\item{min.cells}{Lower limit on the number of identified cells. Defaults to 1000.}

\item{max.cells}{Upper limit on the number of identified cells. Defaults to 15000.}

\item{min.tn5}{Lower threshold for the minimum number of Tn5 integration sites for retaining
a barcode. Defaults to 1000.}

\item{filt.org}{Logical. Whether or not to filter barcodes on based on proportion Tn5 sites occuring
in an organelle. Defaults to FALSE. Expects organelles built off meta}

\item{org.filter.thresh}{Remove cells with an organelle ratio (Organalle/Total_reads) greater than N}

\item{filt.tss}{Logical. Whether or not to filter barcodes on based on proportion Tn5 sites over-
lapping gene TSS. Defaults to TRUE. Filtering barcodes with this parameter is highly recommended.}

\item{tss.min.freq}{Float. Minimum frequency of Tn5 sites near TSSs. Defaults to 0.2.}

\item{tss.z.thresh}{Numeric. Z-score threshold to remove barcodes below the mean. Defeaults to 3
standard deviations (3).}

\item{filt.frip}{Logical. Whether or not to filter barcodes based on the proportion of Tn5 sites over-
lapping bulk ACRs. Defaults to TRUE. Using this filter is highly recommended.}

\item{frip.min.freq}{Float. Minimum frequency of Tn5 sites within ACRs. Defaults to 0.2.}

\item{frip.z.thresh}{Numeric. Z-score threshold to remove barcodes with values below X standard
deviations from the mean. Defaults to 3.}

\item{doplot}{Logical. Whether or not to plot the density scatter plots for tss/frip values.}

\item{prefix}{Character. Prefix output name for plots. If changed from the default (NULL), this will
save the images to disk as a PDF.}
}
\description{
This function filters cells based on read-depth using a mixture of user thresholds and a spline-
fitting algorithm.
}
