% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{generateMatrix}
\alias{generateMatrix}
\title{generateMatrix}
\usage{
generateMatrix(obj, filtered = T, windows = 1000, peaks = FALSE, verbose = T)
}
\arguments{
\item{obj}{Object output from findCells or buildMetaData. Required.}

\item{filtered}{Logical. Whether or not to use the filtered set of cells. Defaults to TRUE.}

\item{windows}{Integer. Window size to build bins. If the 'peaks' parameter is set to TRUE,
this argument is over-ridden.}

\item{peaks}{Logical. If TRUE, use ACRs to build sparse matrix instead of genomic tiles.
Default is set to FALSE.}

\item{verbose}{Logical. Whether or not to print progress.}
}
\description{
This function generates the sparse matrix from equally sized genomic bins or ACRs.
}
