% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccessibility.R
\name{getFDR}
\alias{getFDR}
\title{getFDR}
\usage{
getFDR(obs, exp, fdr = 0.05, grid = 100, verbose = F)
}
\arguments{
\item{fdr}{float. Defaults to 0.05.}

\item{grid}{grid size for determining thresholds. Defaults to 100.}

\item{verbose}{logical. Defaults to TRUE.}

\item{observed}{observed coACRs.}

\item{expected}{shuffled coACRs.}
}
\description{
This function filters coACRs by empirical FDR
}
