% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{loadBEDandGenomeData}
\alias{loadBEDandGenomeData}
\title{loadBEDandGenomeData}
\usage{
loadBEDandGenomeData(bed, ann, sizes, attribute = "Parent", verbose = T)
}
\arguments{
\item{bed}{Path to Tn5 integration site bed file with barcode information.
Required.}

\item{ann}{Path to GFF file for estimating \% reads mapping to TSSs. Required.}

\item{sizes}{Path to chromosome sizes file for downstream processing. Required.}

\item{verbose}{logical. Defaults to TRUE.}
}
\description{
This function loads a 5 column BED file specifying the single-bp Tn5 integration sites.
The file should not have any headers, with the following format:
chromosome, start, end, barcode, strand
}
\details{
The annotation file should be GFF and the chromosome sizes file should be the chromsome ID
followed by its total length (e.g. chr1    123523000) with one chromosome per line.
}
