% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{loadSparseData}
\alias{loadSparseData}
\title{loadSparseData}
\usage{
loadSparseData(input = NULL, meta = NULL, verbose = F)
}
\arguments{
\item{input}{Path to sparse cell x peak text file in triplet format. Input data can be gzipped.
Required.}

\item{meta}{Path to meta data for cells in sparse cell x peak data set in tsv format. Headers
and barcode/cellID rownames are required. Meta data can be gzipped. Providing meta data is
highly recommended, but not required.}

\item{verbose}{logical. Defaults to FALSE.}
}
\description{
This function loads the sparse matrix and meta data and returns a list - counts & meta
}
