% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{logisticModel}
\alias{logisticModel}
\title{Accessibility residuals from unconstrained quasibinomial logistic regression}
\usage{
logisticModel(
  obj,
  r.variates = "~log10nSites",
  variates = NULL,
  nthreads = 1,
  chunks = 256,
  type = "pearson",
  link = "logit",
  center.resid = TRUE,
  scale.resid = FALSE,
  make.sparse = FALSE,
  verbose = FALSE,
  slotName = "residuals"
)
}
\arguments{
\item{obj, }{list containing a slot named 'counts' for the dgCMatrix object with binary cell
(columns) x peak (rows) accessibility scores, and a slot named 'meta' containing a data.frame
with barcode meta data information.}

\item{r.variates}{model formula specifying variable to regress out. Should be specified as
a character string. Defaults to "~log10nSites". Changing this parameter is not recommended.}

\item{variates}{model formula as character string to regress out variables from the initial
residuals using QR residuals in a second round of regression. Ideal for specifying
batch effects, cell-cycle and other technical sources of variation. Defaults to NULL.}

\item{nthreads}{numeric, number of threads to run logistic model in parallel. Depends on
doSNOW. Defaults to 1.}

\item{chunks}{numeric, number of peaks to run for each daughter process. Defaults to 256.}

\item{type}{character, specify the residual type to extract from the model. Defaults to
"pearson".}

\item{link}{character, specify the link-function type for logistic regression. Possible
parameters are "logit" and "probit". Defaults to "logit".}

\item{center.resid}{logical, whether to zero-center residuals. Defaults to TRUE.}

\item{scale.resid}{logical, whther to standardize residuals. Defaults to FALSE.}

\item{make.sparse}{logical, whether or not to set negative values to 0, and reduce memory usage.
Setting this parameter to TRUE has negligible effects on downstream results. Setting make.sparse
to TRUE will override center.resid and scale.resid arguments, setting them to FALSE. Defaults to
FALSE.}

\item{verbose}{logical. Defaults to FALSE.}

\item{slotName}{character, specify the slot name for saving residuals. Useful for saving
multiple normalization steps. Note, make sure to update the slotName argument for
downstream functions. Defaults to "residuals".}
}
\description{
Run unregularized logistic regression. Output is unconstrained Pearson's residuals
for each cell and ACR.
}
