% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{mergeQCdataRDS}
\alias{mergeQCdataRDS}
\title{mergeQCdataRDS}
\usage{
mergeQCdataRDS(filenames)
}
\arguments{
\item{filenames}{Named vector of filepaths.}
}
\description{
This function merges QC data RDS objects. Useful for joining replicates and multiple samples
for downstream analysis. A simple way to aggregate multiple samples into a single socrates
object is to move all the samples of interest into a directory and load the paths with
list.files. IMPORTANT - only merge RDS objects where the same bins/peaks were used for each of
the objects. The easiest way to ensure mergeable objects is to use the same window size (instead
of ACRs) for the function \code{generateMatrix}. Use this function for rds objects saved from
QC analysis. The function \code{mergeSocratesRDS} should be specifically used for objects produced
using \code{convertSparseData}.
}
\section{specify input files}{
input.files <- list.files(pattern="*.rds", path="/path/to/rds/directory")input.files1 "rep1.rds" "rep2.rds"
}

\section{name the input files}{
names(input.files) <- c("rep1", "rep2")input.filesrep1       rep2"rep1.rds" "rep2.rds"
}

\section{merge reps 1 and 2}{
merged.obj <- mergeQCdatRDS(filenames=input.files)
}

\section{check data structure}{
str(merged.obj)
}

