% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{mergeSocratesRDS}
\alias{mergeSocratesRDS}
\title{mergeSocratesRDS}
\usage{
mergeSocratesRDS(filenames = NULL, obj.list = NULL)
}
\arguments{
\item{filenames}{Named vector of filepaths.}

\item{obj.list}{Named list of Socrates objects output from \code{convertSparseData}.}
}
\description{
This function merges Socrates data RDS objects. Useful for joining replicates and multiple samples
for downstream analysis. A simple way to aggregate multiple samples into a single socrates
object is to move all the samples of interest into a directory and load the paths with
list.files. IMPORTANT - only merge RDS objects where the same bins/peaks were used for generating
the sparse matrices. The easiest way to ensure mergeable objects is to use the same window size
(instead of ACRs) for the function \code{generateMatrix} from each sample.
}
\section{specify input files}{
input.files <- list.files(pattern="*.rds", path="/path/to/rds/directory")input.files1 "rep1.rds" "rep2.rds"
}

\section{name the input files}{
names(input.files) <- c("rep1", "rep2")input.filesrep1       rep2"rep1.rds" "rep2.rds"
}

\section{merge reps 1 and 2}{
merged.obj <- mergeSocratesRDS(filenames=input.files)
}

\section{check data structure}{
str(merged.obj)
}

\section{one can also load using a list of socrates objects.}{
all.soc.obj <- list(rep1=soc.obj1, rep22=soc.obj2)merged.obj <- mergeSocratesRDS(obj.list=all.soc.obj)
}

