% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP embeddings}
\usage{
plotUMAP(
  obj,
  column = "LouvainClusters",
  cex = 0.3,
  cluster_slotName = "Clusters",
  xlab = "umap1",
  ylab = "umap2",
  main = "",
  ...
)
}
\arguments{
\item{obj}{list, object containing meta data.}

\item{column}{character, column header specifying how to color cells. Factors are plotted
catagorically, while continuous numeric values are plotted along a spectrum.}

\item{cex}{float, set the point size. Defaults to 0.3.}

\item{cluster_slotName}{character, string specifying the desired UMAP slot to use for plotting.
Deafults to "UMAP".}

\item{xlab}{character string for x-axis name.}

\item{ylab}{character string for y-axis name.}

\item{main}{character string for graph title.}

\item{...}{other arguments accepted by 'plot'.}
}
\description{
Function to plot UMAP embeddings using meta data to establish cell colors.
}
