% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{projectUMAP}
\alias{projectUMAP}
\title{Project cells into a reduced embedding with UMAP}
\usage{
projectUMAP(
  obj,
  m.dist = 0.01,
  k.near = 40,
  metric = "cosine",
  svd_slotName = "PCA",
  umap_slotName = "UMAP",
  verbose = FALSE,
  seed = 1
)
}
\arguments{
\item{obj}{list, object containing 'PCA' for projecting into a reduced embedding with UMAP.}

\item{m.dist}{numeric, m_dist parameter for uwot::umap. Defaults to 0.1}

\item{k.near}{numeric, k-nearest neighbors used by the uwot::umap algorithm. Defaults to 15.}

\item{metric}{character, distance metric used by uwot::umap. Defaults to cosine.}

\item{svd_slotName}{character, name of desired svd_slotName for running UMAP. Defaults to "PCA".}

\item{umap_slotName}{character, name of desired umap_slotName for return UMAP results. Defaults
to "UMAP".}

\item{verbose, }{logical. Defaults to FALSE.}
}
\description{
Project cells into a reduced embedding with UMAP
}
