% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{regModel}
\alias{regModel}
\title{Regularized quasibinomial logistic regression}
\usage{
regModel(
  obj,
  r.variates = "~log10nSites",
  variates = NULL,
  subpeaks = 5000,
  bins = 256,
  bw_adjust = 10,
  type = "pearson",
  link = "logit",
  nthreads = 1,
  weights = NULL,
  method = "lm",
  alpha = 0.5,
  center.resid = T,
  scale.resid = F,
  make.sparse = F,
  verbose = FALSE,
  slotName = "residuals"
)
}
\arguments{
\item{obj, }{list containing a slot named 'counts' for the dgCMatrix object with binary cell
(columns) x peak (rows) accessibility scores, and a slot named 'meta' containing a data.frame
with barcode meta data information.}

\item{r.variates}{model formula specifying variable to regress out. Should be specified as
a character string. Defaults to "~log10nSites". Changing this parameter is not recommended.}

\item{variates}{model formula as character string to regress out variables from the initial
residuals using QR residuals in a second round of regression. Ideal for specifying
batch effects, cell-cycle and other technical sources of variation. Defaults to NULL.}

\item{subpeaks}{numeric, number of ACRs to select for regularization parameter estimates.
Defaults to 5000.}

\item{bins}{numeric, number of bins to split ACRs into. Defaults to 256.}

\item{bw_adjust}{numeric, sets the bandwidth for kernel regression during parameter
regularization. Defaults to 10.}

\item{type}{character, specify the residual type to extract from the model. Possible choices
inlucde: "pearson", "deviance", "response", and "working". Defaults to "pearson".}

\item{link}{character, specify the link-function type for logistic regression. Possible
parameters are "logit" and "probit". Defaults to "logit".}

\item{nthreads}{numeric, number of threads to run logistic model in parallel. Depends on
doSNOW. Defaults to 1.}

\item{weights}{sets weights to glmnet regression to remove technical effects.}

\item{method}{character, type of secondary regression to use for removing additional technical
effects. Choices include "elasticNet" for glmnet based regularized regression or "lm" for
linear regression. Option has no effect when variates is NULL. Using glmnet instead of "lm" is
still experimental and time consuming. Final accessibility residuals should be carefully
analyzed if using "elasticNet". Defaults to "lm".}

\item{alpha}{numeric, sets the alpha parameter for glmnet. 0 for LASSO like regression, 1 for
ridge regression. Default 0.5 for elastic net regression.}

\item{center.resid}{logical, whether to zero-center residuals. Defaults to FALSE.}

\item{scale.resid}{logical, whether to standardize residuals. Defaults to TRUE.}

\item{make.sparse}{logical, whether or not to set negative values to 0, and reduce memory usage.
Setting this parameter to TRUE has negligible effects on downstream results. Setting make.sparse
to TRUE (default) will override center.resid and scale.resid arguments, setting them to FALSE. make.sparse
is mutally exclusive with the argument 'variates'. Defaults to FALSE}

\item{verbose}{logical. Defaults to FALSE.}

\item{slotName}{character, specify the slot name for saving residuals. Useful for saving
multiple normalization steps. Note, make sure to update the slotName argument for
downstream functions. Defaults to "residuals".}
}
\description{
Regularized quasibinomial logistic regression
}
