% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaccessibility.R
\name{runCicero}
\alias{runCicero}
\title{runCicero}
\usage{
runCicero(
  cds,
  genome = NULL,
  k = 30,
  win = 5e+05,
  sample_num = 100,
  silent = TRUE
)
}
\arguments{
\item{cds}{cicero object. Required.}

\item{genome}{path to genome file. See 'inst/extdata/genome' for an example. Required.}

\item{k}{numeric. Number of nearest neighbors for binning cells. See cicero for more details.
Defaults to 30.}

\item{win}{numeric. Window size for constraining co-accessible ACRs. Defaults to 500000.}

\item{sample_num}{numeric. Number of random windows to sample for determining distance_parameter
during 'run_cicero'. See cicero for more details.}

\item{silent}{logical. Defaults to TRUE.}
}
\description{
This function is a wrapper for several cicero commands
}
