% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{standardizeResiduals}
\alias{standardizeResiduals}
\title{standardize or center residuals}
\usage{
standardizeResiduals(obj, center = T, scale = F, bins = 1024, verbose = F)
}
\arguments{
\item{obj}{list object containing slot 'residuals' output from logisticModel or regModel functions.}

\item{center}{logical, whether or not to center residuals. Has no effect if scale is TRUE.
Defaults to TRUE.}

\item{scale}{logical, whether or not to center and standardize residuals. Defaults to TRUE.}

\item{bins}{numeric, number of bins to split ACRs into. Defaults to 1024.}

\item{verbose}{logical. Defaults to FALSE.}
}
\description{
iterate over residual matrix, scaling or centering values across cells for each ACR.
}
