% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{writeReadME}
\alias{writeReadME}
\title{writeReadME}
\usage{
writeReadME(fileID = "QC_README")
}
\arguments{
\item{fileID}{Character string specifying the output file name.}
}
\description{
This function generates a README of the output files from quality control
}
