% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublet_analysis.R
\name{detectDoublets}
\alias{detectDoublets}
\title{detectDoublets}
\usage{
detectDoublets(
  obj = NULL,
  nTrials = 5,
  nSample = 1000,
  k = 10,
  n.pcs = 50,
  seed = 1,
  threads = 1,
  svd_slot = "PCA"
)
}
\arguments{
\item{obj}{Socrates object. For best results, use downstream of the function cleanCells prior to
normalization.}

\item{nTrials}{Numeric. Number of times nSample cells are simulated.}

\item{nSample}{Numeric. Number of synthetic doublets to create per trial.}

\item{k}{Numeric. Number of nearest neighbors to search for when estimating doublet enrichment via knn
from the narbor package.}

\item{n.pcs}{Numeric. Number of PCS/SVD components to retain when reducing dimensions.}

\item{seed}{Numeric.}

\item{threads}{Numeric. Number of threads to use for mclapply from the parallel package.}

\item{normModel}{Character string. normalization model for embedding synthetic doublets. Options include
'regModel', 'regModel2', 'logisticModel', and 'tfidf'. TF-IDF is the default due to its quicker run time.
But the other models are also available if the user wish to maintain a consistent normalization scheme
as the intended analysis.}
}
\description{
This function estimates doublet likelihood via synthetic doublet creation and enrichment
analysis. The underlying code is heavily adapted from ArchR described by Granja and Corces et al.
If you use detectDoublets in your analysis, please cite the original ArchR paper.
}
