% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{reduceDims}
\alias{reduceDims}
\title{Reduce dimensions of residual matrix}
\usage{
reduceDims(
  obj,
  n.pcs = 50,
  scaleVar = T,
  cor.max = 0.75,
  doL2 = NULL,
  stdLSI = 1,
  residuals_slotName = "residuals",
  svd_slotName = "PCA",
  verbose = FALSE
)
}
\arguments{
\item{obj}{list, object containing "pearson_residuals" output by the function regModel.}

\item{n.pcs}{numeric, number of singular values to calculate.}

\item{cor.max}{float, maximum spearman correlation between log10nSites (log10 number of
accessible peaks) and singular value to keep. Singular values with correlations greater than
cor.max are removed. Ranges from 0 to 1. Default set to 0.75.}

\item{doL2}{numeric, whether to L2 normalize barcodes (doL2=1), L2 normalize PCs (doL2=2),
or no L2 normalization (NULL or FALSE).}

\item{stdLSI}{numeric, whether to standarize barcodes (stdLSI=1), standardize PCs (stdLSI=2),
or no SVD standarization (NULL or FALSE). Defaults to 1 (standardize components for each
barcode).}

\item{residuals_slotName}{character, character string of the desired residual slotName. Defaults
to "residuals".}

\item{svd_slotName}{character, character string for naming the SVD output in the returned
object. Defaults to "PCA".}

\item{verbose}{logical. Defaults to FALSE.}

\item{scaleVars}{logical, whether or not to scale PCs by variance explained. Default to TRUE.}
}
\description{
Run SVD on Pearson residuals using IRLBA.
}
