% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{reduceDims}
\alias{reduceDims}
\title{Reduce dimensions of residual matrix}
\usage{
reduceDims(
  obj,
  method = "SVD",
  n.pcs = 50,
  scaleVar = T,
  num.var = 5000,
  cor.max = 0.75,
  doL2 = F,
  doL1 = F,
  doSTD = T,
  residuals_slotName = "residuals",
  svd_slotName = "PCA",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{obj}{list, object containing "pearson_residuals" output by the function regModel.}

\item{method}{character, string denoting dimension reduction method. Can be one of "SVD" or
"NMF". Defaults to SVD.}

\item{n.pcs}{numeric, number of singular values to calculate.}

\item{cor.max}{float, maximum spearman correlation between log10nSites (log10 number of
accessible peaks) and singular value to keep. Singular values with correlations greater than
cor.max are removed. Ranges from 0 to 1. Default set to 0.75.}

\item{doL2}{logical, whether or not to L2 normalize barcodes.}

\item{doL1}{logical, whether or not to L1 normalize barcodes}

\item{residuals_slotName}{character, character string of the desired residual slotName. Defaults
to "residuals".}

\item{svd_slotName}{character, character string for naming the SVD output in the returned
object. Defaults to "PCA".}

\item{verbose}{logical. Defaults to FALSE.}

\item{...}{Additional arguments to be passed to RcppML::nmf}

\item{scaleVars}{logical, whether or not to scale PCs by variance explained (or to scale
NMF components by scale factors). Default to TRUE.}

\item{num.vars}{number of highly variable ACRs/bins to use for dimensionality reduction.
Defaults to 5000. Set to NULL to use all ACRs/bins.}

\item{stdLSI}{logical, whether or not to standardize barcodes.}
}
\description{
Run SVD on Pearson residuals using IRLBA.
}
